Fix Safari Extension code bug
The Safari Extension has dedicated code within the browser, that is not part of the standard JavaScript. When using this code, it is minified because of no externs data for the extension.
As of now, I've just made a quick fix such that it works, but in time I'd like to add a correct externs file to the script.